remove client-closure timeouts
Some users were experiencing bugs with the client closure timeout.

I don't think there is a great need to complicate the program with the
client closure timeout. It may be that the backend experiences more
connections than usual but only because it is slow in interpretting the
failure and nginx is fast. Since this is an exceptional situation to begin
with coding around probably isn't warrented.